+++ anyblok/registry.py
@@ -80,7 +80,6 @@ class RegistryManager:
     callback_initialize_entries = {}
     callback_unload_entries = {}
     registries = {}
-    needed_bloks = []
 
     @classmethod
     def has_blok(cls, blok):
@@ -404,16 +403,6 @@ class RegistryManager:
         if cls.has_blok_property(property_):
             del cls.loaded_bloks[blok]['properties'][property_]
 
-    @classmethod
-    def add_needed_bloks(cls, *bloks):
-        for blok in bloks:
-            if blok not in cls.needed_bloks:
-                cls.needed_bloks.append(blok)
-
-    @classmethod
-    def get_needed_bloks(cls):
-        return cls.needed_bloks
-
 
 class Registry:
     """ Define one registry
@@ -580,10 +569,6 @@ class Registry:
         """
         toinstall = self.get_bloks_by_states('toinstall')
         for blok in BlokManager.auto_install:
-            if blok not in (toinstall + loaded):
-                toinstall.append(blok)
-
-        for blok in RegistryManager.get_needed_bloks():
             if blok not in (toinstall + loaded):
                 toinstall.append(blok)